Align Open API specs, Entity and DTO numeric types with DB sizes to avoid excessive conversions. (#2032)

+++ src/main/java/uk/gov/hmcts/opal/dto/CourtReferenceDto.java
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 public class CourtReferenceDto implements ToJsonString {
 
     @JsonProperty("court_id")
-    private Integer courtId;
+    private Long courtId;
 
     @JsonProperty("court_name")
     private String courtName;
